+++ src/Accessor.ts
@@ -68,15 +68,12 @@ export class Getter implements Accessor {
 
     async definition(target: SourceDocument, position: vscode.Position, curlySeparator: string): Promise<string> {
         const eol = target.endOfLine;
-        const templateStatement = this.memberVariable.parent?.isUnspecializedTemplate()
-                ? this.memberVariable.parent.templateStatement(true) + eol
-                : '';
         const inlineSpecifier =
             ((!this.parent || !util.containsExclusive(this.parent.range, position))
             && this.memberVariable.document.fileName === target.fileName)
                 ? 'inline '
                 : '';
-        return templateStatement + inlineSpecifier + this.returnType
+        return this.memberVariable.combinedTemplateStatements(true, eol) + inlineSpecifier + this.returnType
                 + await this.memberVariable.scopeString(target, position) + this.name + '()'
                 + (this.isStatic ? '' : ' const') + curlySeparator + '{'
                 + eol + util.indentation() + this.body + eol + '}';
@@ -136,15 +133,12 @@ export class Setter implements Accessor {
 
     async definition(target: SourceDocument, position: vscode.Position, curlySeparator: string): Promise<string> {
         const eol = target.endOfLine;
-        const templateStatement = this.memberVariable.parent?.isUnspecializedTemplate()
-                ? this.memberVariable.parent.templateStatement(true) + eol
-                : '';
         const inlineSpecifier =
             ((!this.parent || !util.containsExclusive(this.parent.range, position))
             && this.memberVariable.document.fileName === target.fileName)
                 ? 'inline '
                 : '';
-        return templateStatement + inlineSpecifier + this.returnType
+        return this.memberVariable.combinedTemplateStatements(true, eol) + inlineSpecifier + this.returnType
                 + await this.memberVariable.scopeString(target, position) + this.name + '(' + this.parameter + ')'
                 + curlySeparator + '{' + eol + util.indentation() + this.body + eol + '}';
     }