+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to the `C-mantic` extension will be documented in this file.
 ## [Unreleased]
 ### Added
 - Added support for explicit template specialization. Definitions for specialized function templates/member functions of specialized class templates can now be generated in the matching source file.
+- Added support for nested templates.
 
 ### Fixed
 - Fixed title of `Add Declaration` code-action displaying the wrong target file.